PATRICIO, Chrisjane
The code you provided seems to be mostly correct. However, there is one error and one improvement you can make:

Error: Forgot to provide space in the

import 'package: flutter/material.dart';

To fix this, add space between import and the ('). 
Improvement: The improvement can make is to wrap the Container widget inside an Align widget instead of using the Center widget. This change will allow to align the container within the available space. 